requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
gensim==4.3.1
torch==2.0.1
transformers==4.31.0
tensorflow-hub==0.14.0
tensorflow==2.13.0
flask==2.3.3
mlflow==2.6.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
pyLDAvis==3.4.1
evidently==0.4.0
prometheus-client==0.17.1
kedro==0.18.14
pytest==7.4.0
python-multipart==0.0.6
gunicorn==21.2.0 